requirements.txt
langchain
langchain-ollama
mysql-connector-python
tqdm
pandas
sentence-transformers
qdrant-client